Refactor calendar form - Move all logic into custom hook - Break up into smaller components
